Address workflow reviews and clean up stage preparation
