+++ skills/applications/tao-run-deft-cr-its-mining/scripts/log_stage.py
@@ -20,17 +20,16 @@ from workflow_common import atomic_write_json
 VALID_STATUSES = {"ok", "error", "skipped"}
 VALID_STAGES = {
     "validate_workflow",
-    "init_state",
+    "initialize_workflow",
     "baseline_evaluate",
-    "setup_embeddings",
+    "prepare_cosmos_embed_inference",
     "cosmos_embed",
     "convert_embeddings",
     "gap_analysis",
-    "build_mining_target",
+    "prepare_nearest_neighbor_mining",
     "mine_nearest_neighbors",
     "record_mined_paths",
-    "build_llava_from_mining",
-    "assemble_train_annotations",
+    "prepare_cosmos_reason_train",
     "train",
     "evaluate",
     "loop_stop",
